Cassandra/requirements.txt
cassandra-driver>=3.13.0
msgpack-python>=0.5.5
